### 2021-09-23 #3757 @whedon
[PRE REVIEW]: POWERsim: A nanoPOWER ISA Functional Simulator
**Submitting author:** @god-s-perfect-idiot (<a href="http://orcid.org/0000-0003-2213-8820">Samar Musthafa</a>)
**Repository:** <a href="https://github.com/god-s-perfect-idiot/POWER-sim" target ="_blank">https://github.com/god-s-perfect-idiot/POWER-sim</a>
**Version:** v0.7x
