### 2021-09-23 #3757 @kthyng
"Can you edit your paper to begin with a jargon-free explanation of the purpose of the software that a technical person outside of your field can understand?" was meant to refer to the summary and "Also, the statement of need should explain why the software is useful (less jargon and fewer acronyms is always better) as well as the research applications." to the statement of need section.
