### 2021-09-23 #3757 @kthyng
@god-s-perfect-idiot I don't see improvement in the paper, just a few small changes.  Here are more suggestions:

Summary:
* what is "instruction set architecture" and why do people care about it?
* what field or fields is this in?

Statement of need:
* Who might use this software
* What software already exists in this area
* Compare this software with the software that already exists
* Make a compelling argument for this being research software, as defined in [our docs](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software)

Please read carefully through our guidelines for what should be in your [paper](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain).
