+++ paper.bib
@@ -4,13 +4,30 @@
        title = {Power ISA Manualv3.0B},
        howpublished = "\url{https://ibm.ent.box.com/s/1hzcwkwf8rbju5h9iyf44wm94amnlcrv}"
 }
+@misc{isa,
+       author = {\textbf{Wikipedia}},
+       title = {Instruction Set Architecture},
+       howpublished = "\url{https://en.wikipedia.org/wiki/Instruction_set_architecture#:~:text=In%20computer%20science%2C%20an%20instruction,)%2C%20is%20called%20an%20implementation.}"
+}
+@misc{power10fs,
+       author = {\textbf{IBM}},
+       title = {Power10 Functional Simulator},
+       howpublished = "\url{https://openpowerfoundation.org/introducing-ibm-power10-functional-simulator/}"
+}
+@misc{microwatt, 
+  title={Microwatt},
+  author={\textbf{Anton Blanchard}}, 
+  year={2021}, 
+  publisher={Github},
+  journal = {GitHub repository},
+  url = {https://github.com/antonblanchard/microwatt}
+}
 @misc{pythonlibs, 
   title={The Python Library Reference, release 3.8.2},
   author={\textbf{Van Rossum, Guido}}, 
   year={2021}, 
   publisher={Python Software Foundation} 
 }
-
 @misc{cx_Freeze,
   author = {Anthony Tuininga},
   title = {cx_Freeze},
