### 2021-10-14 #3757 @god-s-perfect-idiot
Hi @jarvist thank you for taking up the editing. I was hoping @jedbrown could take a look as i felt they had the best profile and background for an ISA simulator. But, im not sure if they're available, or if im incorrect on said notion.
