### 2021-12-03 #3973 @whedon
[REVIEW]: POWERsim: A nanoPOWER ISA Functional Simulator
**Submitting author:** @god-s-perfect-idiot (<a href="http://orcid.org/0000-0003-2213-8820">Samar Musthafa</a>)
**Repository:** https://github.com/god-s-perfect-idiot/POWER-sim
**Branch with paper.md** (empty if default branch):
**Version:** v0.7x
