### 2021-12-03 #3973 @jarvist
Thank you again @dilawar and @federeghe for agreeing to review this paper & package.

Your task lists are in the first note on this issue, along with some links to the JOSS guidelines. A key thing I often accidentally do when reviewing is start to check off the other reviewer's list.

Would a deadline of about 2 weeks work for the first review? Otherwise I think we will start getting caught up in the Christmas / New Year closure. One of the nice things about JOSS is that a partial review is still really useful, and add comments here / issues on the package itself to give @god-s-perfect-idiot formative feedback.
