### 2021-12-03 god-s-perfect-idiot/POWER-sim#3 @dilawar
Can't get started, missing documentation and basic hygiene  
Related: https://github.com/openjournals/joss-reviews/issues/3973#issuecomment-985644450

I am following README.md file, which is broken and confusing.

I did the following after cloning your repository.

```bash
[dilawars@chutki POWER-sim (master)]$ python launchers/cli.py
Traceback (most recent call last):
  File "/home/dilawars/Work/GITHUB/JOSS/POWER-sim/launchers/cli.py", line 9, in <module>
    import assembler as asm
ModuleNotFoundError: No module named 'assembler'
```
This is what I thought a normal user would do after reading
> CLI Mode: run: `launchers/cli.py -m operation mode -i input file -o output file` input and output files are optional

I see that there is `asm-proc/assembler.py` so I am assuming this is the module which `cli.py` is trying to import? I could overcome the `ModuleNotFoundErrror` by doing `cd launchers && python cli.py`. Now, I don't know available values for `-m/--mode`? I could not find anything in documentation.md or README.md. Note that casual users are not expected to read the source code to find these values.

Most Python users would expect a standard project structure. It would help if you can reorganize your project structure to some standard. Here is a good option: https://docs.python-guide.org/writing/structure/. Otherwise, it's going to cause unnecessary confusion.

Ideally, add a CI/CD pipeline that builds your application and run tests. Github has [Actions](https://github.com/features/actions) that are easy to follow. You can also add some other CI/CD pipeline. In fact, a simple bash script or  Makefile would also be fine.

Also add a dependency file: `requirements.txt` is a good option, `setup.py` is another.

I'd have liked `python setup.py install` to install the project or even better `pip install POWER_sim` installing from pypi.org.

At this stage of review, this project would require a clearer documentation, test suite, API documentation (doc strings are non-existent) and optionally a CI/CD pipeline (or some other automation to run the test-suite).
