### 2021-12-14 god-s-perfect-idiot/POWER-sim#4 @federeghe
README and launch problems
Related: openjournals/joss-reviews#3973

Similarly to the comments of the other JOSS reviewer, I cannot start the software with the error:
```
ModuleNotFoundError: No module named 'assembler'
```
I can start the software if my working directory is `launchers`. This highlights some problems in the management of the Python modules and directories.

In addition:
- Missing system `patchelf` dependency in the README
- pip package incorrectly spelled in the README, should be`cx_Freeze` not `csx_Freeze`
- Missing [hashbangs](https://en.wikipedia.org/wiki/Shebang_(Unix)) in the python scripts that may lead to run `python2` instead of `python3`

The README is not clear and clean, please add some description of the software in the first lines, at least which architectures support. I would expect to find this as a user as the first information. Also, I would move the launch modes after the installation (considering also the windows-only command).
