### 2021-12-14 god-s-perfect-idiot/POWER-sim#5 @federeghe
Improve input/output handling in CLI script and documentation
Related: openjournals/joss-reviews#3973

The CLI help does not really provide many useful information:
```
optional arguments:
  -h, --help            show this help message and exit
  -i INPUT, --input INPUT
                        Input file
  -o OUTPUT, --output OUTPUT
                        Output file
  -m MODE, --mode MODE  POWERsim operational mode
```

- I expect it to print a detailed help (what is the input file, what is the output, which modes are available, etc.)
- What values are possible for mode? The README says "Assembler Mode" and "Processor Mode":
  - How do I specify these modes? "Assembler"? "assembler"? "a"?
  - The cli.py seems to accept any string: `python3 cli.py -m hello_world` works, not sure what it does
  - The difference between the two modes explained by the README is not clear to me, please improve it
- How is possible that input and output are optional? What does the ISA simulator run without an input file? This should be explained
- The script seems to accept a non-existent input file without any warning, not sure what it does
