### 2021-12-15 #3973 @jarvist
@dilawar , how are things coming along?
I appreciate things might be getting blown off course by this latest Omicron situation!
