### 2021-12-15 #3973 @dilawar
I have opened a ticket on their repo but haven't heard back from authors yet.

Unless they fix it there is no way to run the tool and review its capabilities. I am hoping the author will respond soon and will be more proactive in the future.

So far things haven't become any worse here due to Omicron (I guess its same for the author).
Sent from a handheld device, pardon the brevity and tpyos.

15-Dec-2021 20:20:07 Jarvist Moore Frost ***@***.***>:

> @dilawar[https://github.com/dilawar] , how are things coming along?
> I appreciate things might be getting blown off course by this latest Omicron situation!
>
