### 2021-12-15 #3973 @jarvist
Thank you @dilawar  - I missed this feedback. That is indeed a bit of a show stopper. I see also that @federeghe  is having similar problems.

@god-s-perfect-idiot is there some timeline on which you would be able to provide a conda environment / pip install / setup.py etc. for more automated installation?
