### 2022-01-07 #3973 @jarvist
Dear @god-s-perfect-idiot , do you have a timescale / plan for resolving issues https://github.com/god-s-perfect-idiot/POWER-sim/issues/4 and https://github.com/god-s-perfect-idiot/POWER-sim/issues/3 so that the reviewers are able to review your software?
